+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased changes
 
+- Make fields of CIS4 events public.
+
 ## 4.0.0 (2024-01-22)
 
 - Add `MultiExp` trait that allows to have different `multiexp` algorithm implementations for different curves.