+++ rust-src/concordium_base/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased changes
 
+- Set minimum supported Rust version to 1.73.
+- Set `chrono` to version `0.4.35`.
+- Make fields of CIS4 events public.
+
 ## 4.0.0 (2024-01-22)
 
 - Add `MultiExp` trait that allows to have different `multiexp` algorithm implementations for different curves.