+++ rust-src/concordium_base/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 - Set minimum supported Rust version to 1.73.
 - Make fields of CIS4 events public.
+- Remove the `From<SlotDuration>` and `From<DurationSeconds>` implementations
+  for `chrono::Duration` and replace them with fallible `TryFrom`
+  implementations that fail when durations overflow.
+- `ContractAddress::new`
+- `ContractName`, `ReceiveName`, `EntrypointName`, and `Parameter`
+  `new_unchecked` constructors are made `const` so they can be used when
+  defining constants. Similarly `Parameter::empty` is `const` now.
 
 ## 4.0.0 (2024-01-22)